fix(ui): address #724 review — capture active group before await; dedupe thunk-test harness
- enumeration.thunks: read selectActiveGroupId before the POST await so a group switch
  mid-request can't retarget the list refetch (Greptile P2). Keep active group (not the
  new dataset's group) so the refetch matches the Datasets page view, incl. "All Groups".
- Extract the duplicated recording-store thunk-test harness into test/recordingStore.ts and
  use it in the enumeration/reactions/datasets thunk tests (clears SonarCloud new-code
  duplication, which flagged the copy-pasted makeStore in the new test).
